+++ sample/src/main/java/com/doodle/android/chips/sample/MainActivity.java
@@ -16,8 +16,10 @@
 
 package com.doodle.android.chips.sample;
 
+import android.graphics.Typeface;
 import android.net.Uri;
 import android.os.Bundle;
+import android.support.v4.app.FragmentManager;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
@@ -54,6 +56,9 @@ public class MainActivity extends AppCompatActivity {
 
         mChipsView = (ChipsView) findViewById(R.id.cv_contacts);
 
+        mChipsView.setTypeface(Typeface.createFromAsset(this.getAssets(), "fonts/FiraSans-Medium.ttf"));
+        // mChipsView.useInitials(14, Typeface.createFromAsset(this.getAssets(), "fonts/FiraSans-Medium.ttf"), Color.RED);
+
         // change EditText config
         mChipsView.getEditText().setCursorVisible(true);
 
@@ -83,6 +88,36 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onTextChanged(CharSequence text) {
                 mAdapter.filterItems(text);
+            }
+
+            @Override
+            public boolean onInputNotRecognized(String text) {
+
+                try {
+                    FragmentManager fragmentManager = MainActivity.this.getSupportFragmentManager();
+
+                    Bundle bundle = new Bundle();
+                    bundle.putString(ChipsEmailDialogFragment.EXTRA_STRING_TEXT, text);
+                    bundle.putString(ChipsEmailDialogFragment.EXTRA_STRING_TITLE, "Title");
+                    bundle.putString(ChipsEmailDialogFragment.EXTRA_STRING_PLACEHOLDER, "ChipsDialogPlaceholder");
+                    bundle.putString(ChipsEmailDialogFragment.EXTRA_STRING_CONFIRM, "ChipsDialogConfirm");
+                    bundle.putString(ChipsEmailDialogFragment.EXTRA_STRING_CANCEL, "ChipsDialogCancel");
+                    bundle.putString(ChipsEmailDialogFragment.EXTRA_STRING_ERROR_MSG, "ChipsDialogErrorMsg");
+
+                    ChipsEmailDialogFragment chipsEmailDialogFragment = new ChipsEmailDialogFragment();
+                    chipsEmailDialogFragment.setArguments(bundle);
+                    chipsEmailDialogFragment.setEmailListener(new ChipsEmailDialogFragment.EmailListener() {
+                        @Override
+                        public void onDialogEmailEntered(String text, String displayName) {
+                            mChipsView.addChip(displayName, null, new Contact(null, null, displayName, text, null), false);
+                            mChipsView.clearText();
+                        }
+                    });
+                    chipsEmailDialogFragment.show(fragmentManager, ChipsEmailDialogFragment.class.getSimpleName());
+                } catch (ClassCastException e) {
+                    Log.e("CHIPS", "Error ClassCast", e);
+                }
+                return false;
             }
         });
     }
@@ -166,7 +201,8 @@ public class MainActivity extends AppCompatActivity {
             Contact contact = new Contact(null, null, null, email, imgUrl);
 
             if (selection.isChecked()) {
-                mChipsView.addChip(email, imgUrl, contact);
+                boolean indelibe = Math.random() > 0.8f;
+                mChipsView.addChip(email, imgUrl, contact, indelibe);
             } else {
                 mChipsView.removeChipBy(contact);
             }